+++ ling_class/genxml.cc
@@ -254,8 +254,8 @@ static void extract_ids(XML_Attribute_List &attributes,
   else if (attributes.present("href"))
     {
       val = attributes.val("href");
-      int starts[EST_Regex_max_subexpressions];
-      int ends[EST_Regex_max_subexpressions];
+      size_t starts[EST_Regex_max_subexpressions];
+      size_t ends[EST_Regex_max_subexpressions];
       
       if (val.matches(simpleIDRegex, 0, starts, ends))
 	{
@@ -418,11 +418,11 @@ static void proccess_features(EST_String name,
 			     EST_Features &f)
 {
   EST_String names[MAX_FEATS];
-  int starts[EST_Regex_max_subexpressions];
-  int ends[EST_Regex_max_subexpressions];
+  size_t starts[EST_Regex_max_subexpressions];
+  size_t ends[EST_Regex_max_subexpressions];
   
-  int n = split(defs, names, MAX_FEATS, feat_sep);
-  for(int i=0; i<n; i++)
+  size_t n = split(defs, names, MAX_FEATS, feat_sep);
+  for(size_t i=0; i<n; i++)
     {
       EST_String def = names[i];
       EST_String feat;
